+++ lib/puree/configuration.rb
@@ -1,0 +1,17 @@
+module Puree
+
+  module Configuration
+
+    class << self
+
+      attr_accessor :endpoint, :username, :password
+
+      def configure
+        yield self
+      end
+
+    end
+
+  end
+
+end