+++ lib/puree/resource.rb
@@ -4,28 +4,38 @@ module Puree
   #
   class Resource
 
-    def initialize(resource_type)
-      @resource_type = resource_type
+    # @param api [String]
+    # @param endpoint [String]
+    # @param optional username [String]
+    # @param optional password [String]
+    def initialize( api: nil,
+                    endpoint: nil,
+                    username: nil,
+                    password: nil)
+      @resource_type = api
       @api_map = Puree::Map.new.get
+      @endpoint = endpoint.nil? ? Puree::Configuration.endpoint : endpoint
+      @username = username.nil? ? Puree::Configuration.username : username
+      @password = password.nil? ? Puree::Configuration.password : password
     end
 
     # Get
     #
-    # @param endpoint [String]
-    # @param username [String]
-    # @param password [String]
     # @param uuid [String]
     # @param id [String]
     # @return [HTTParty::Response]
-    def get(endpoint:       nil,
-            username:       nil,
-            password:       nil,
-            uuid:           nil,
-            id:             nil
-    )
+    def get(uuid: nil, id: nil)
+      missing = missing_credentials
+      if !missing.empty?
+        missing.each do |m|
+          puts "#{self.class.name}" + '#' + "#{__method__} missing #{m}"
+        end
+        exit
+      end
+
       # strip any trailing slash
-      @endpoint = endpoint.sub(/(\/)+$/, '')
-      @auth = Base64::strict_encode64(username + ':' + password)
+      @endpoint = @endpoint.sub(/(\/)+$/, '')
+      @auth = Base64::strict_encode64(@username + ':' + @password)
 
       @options = {
           latest_api:     true,
@@ -124,7 +134,18 @@ module Puree
       o
     end
 
+
+
     private
+
+
+    def collect_uuid
+      path = '//renderedItem/@renderedContentUUID'
+      xpath_result = xpath_query path
+      xpath_result.each { |i| @uuids << i.text.strip }
+    end
+
+
 
 
 
@@ -171,6 +192,21 @@ module Puree
       doc.xpath path
     end
 
+    def missing_credentials
+      missing = []
+      if @endpoint.nil?
+        missing << 'endpoint'
+      end
+      if @username.nil?
+        missing << 'username'
+      end
+      if @password.nil?
+        missing << 'password'
+      end
+      missing
+    end
+
+    alias :find :get
 
   end
 end